Move CLI offline args to clap-utils
