+++ cli/src/offline/blockhash_query.rs
@@ -1,5 +1,10 @@
 use super::*;
-use solana_clap_utils::nonce::*;
+use clap::ArgMatches;
+use solana_clap_utils::{
+    input_parsers::{pubkey_of, value_of},
+    nonce::*,
+    offline::*,
+};
 use solana_client::nonce_utils;
 use solana_sdk::commitment_config::CommitmentConfig;
 
@@ -110,7 +115,7 @@ impl Default for BlockhashQuery {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::offline::blockhash_query::BlockhashQuery;
+    use crate::offline::blockhash_query::{self, BlockhashQuery};
     use clap::App;
     use serde_json::{self, json, Value};
     use solana_account_decoder::{UiAccount, UiAccountEncoding};
